+++ PHP/JSON-RPC/example_local.php
@@ -6,7 +6,9 @@ include(dirname(__FILE__) . '/SelectSpecs_v_0_7/include.php');
 // we can use custom parameters for connection directly in code (NOT Recomended), please use config.php
 //$ssAPI = new SSAPI('api.example.com/json-rpc', 8843, 'secret token', 'group id');
 try {
-    $ssapi = new SSAPI("127.0.0.1", 8843, "SeCrEtToKeNvAlUe", "GrOuPiD");
+    //111c1111111111cc11111111 - test group_id
+    //SeCrEtToKeNvAlUe - test token
+    $ssapi = new SSAPI("127.0.0.1", 8843, "SeCrEtToKeNvAlUe", "111c1111111111cc11111111");
 } catch (Exception $e) {
     echo $e;
     die('<br><br><hr><b>CANT CONNECT TO SERVER');
@@ -14,40 +16,53 @@ try {
 
 
 //Select orders by query
-    $result = $ssapi->orders(['id' => 1]);
+    echo "<b>SELECT:</b>\n";
+    $result = $ssapi->orders(['order_id' => 4]);
     var_dump($result);
 
     profiling();
 
-
-//Update orders by query
-    $result = $ssapi->orders(['id' => 2], ['name' => 'test order 2', '_group_id' => 'A1498Hjhjh99hjjh'], SSAPI_RETURN_RESULT);
+//Remove orders by query
+    echo "<b>Remove:</b>\n";
+    $result = $ssapi->orders(['order_id' => 4], NULL, SSAPI_DELETE_DOCUMENT);
     var_dump($result);
 
     profiling();
+
 
 //Replace orders by query
-    $result = $ssapi->orders(['id' => 3], ['name' => 'test order 3', '_group_id' => 'A1498Hjhjh99hjjh'], SSAPI_CREATE_IF_NOT_EXIST | SSAPI_RETURN_RESULT);
+    echo "<b>REPLACE (with return result flag):</b>\n";
+    $result = $ssapi->orders(['order_id' => 3], ['name' => 'test order 3', '_group_id' => '333d3333333333dd33333333'], SSAPI_CREATE_IF_NOT_EXIST | SSAPI_RETURN_RESULT);
     var_dump($result);
 
     profiling();
 
 //Insert order
-    $result = $ssapi->orders(NULL, ['name' => 'test order 4', '_group_id' => 'A1498Hjhjh99hjjh'], SSAPI_CREATE_IF_NOT_EXIST | SSAPI_RETURN_RESULT);
+    echo "<b>INSERT (with return result flag):</b>\n";
+    $result = $ssapi->orders(NULL, ['order_id' => 4, 'name' => 'test order 4', '_group_id' => '333d3333333333dd33333333'], SSAPI_RETURN_RESULT);
+    var_dump($result);
+
+    profiling();
+
+//Update orders by query
+    echo "<b>UPDATE (with return result flag):</b>\n";
+    $result = $ssapi->orders(['order_id' => 4, '_group_id' => '333d3333333333dd33333333'], ['name' => 'UPDATED order 4 to 2', '_group_id' => '444d4444444444dd44444444'], SSAPI_RETURN_RESULT);
     var_dump($result);
 
     profiling();
 
 //Insert order and don't wait answer from server
-    $ssapi->orders(NULL, ['name' => 'test order 5', '_group_id' => 'A1498Hjhjh99hjjh'], SSAPI_NO_WAIT_RESPONSE);
-    var_dump('NOTIFICATION');
+    $ssapi->orders(NULL, ['order_id' => 5, 'name' => 'test order 5', '_group_id' => '333d3333333333dd33333333'], SSAPI_NO_WAIT_RESPONSE);
+    echo "INSERT (w/o answer)\n";
+
     profiling();
 
 //Replace order and don't wait answer from server
-    $ssapi->orders(['id' => 6], ['name' => 'test order 6', '_group_id' => 'A1498Hjhjh99hjjh'], SSAPI_CREATE_IF_NOT_EXIST | SSAPI_NO_WAIT_RESPONSE);
-    var_dump('NOTIFICATION');
+    $ssapi->orders(['order_id' => 6], ['name' => 'test order 6', '_group_id' => '333d3333333333dd33333333'], SSAPI_CREATE_IF_NOT_EXIST | SSAPI_NO_WAIT_RESPONSE);
+    echo "REPLACE (w/o answer)\n";
 
     profiling();
+*/
     profiling(true);