+++ PHP/JSON-RPC/Docs/changelog.md
@@ -1,5 +1,6 @@
 ## version 0.7.0
 - methods names separation added (IMPORTANT UPDATE! previus versions is will no work anymore)
+- examples was fixed (_group_id now have right format)
 
 ## version 0.6.2
 - fixed constructor defaults declaration for PHP 5.5